oracle-install-test-v3

+++ main.tf
@@ -39,19 +39,29 @@ resource "google_compute_instance" "default" {
 
   metadata_startup_script = <<-EOF
     #!/bin/bash
-    #sudo yum update -y #save time for first run
-    sudo yum install wget -y
-    sudo mkdir sw
-    sudo mkdir log
-    chmod 777 sw
-    chmod 777 log
-    cd sw
-    wget https://yum.oracle.com/repo/OracleLinux/OL9/appstream/x86_64/getPackage/oracle-database-preinstall-23ai-1.0-2.el9.x86_64.rpm
-    wget https://download.oracle.com/otn-pub/otn_software/db-free/oracle-database-free-23ai-1.0-1.el9.x86_64.rpm
-    sudo dnf -y install oracle-database-preinstall-23ai-1.0-2.el9.x86_64.rpm > /home/peter/log/install_log.txt
-    sudo dnf install -y oracle-database-free* >> /home/peter/log/install_log.txt
-    #export DB_PASSWORD=sys
-    (echo sys; echo sys;) | sudo /etc/init.d/oracle-free-23ai configure >> /home/peter/log/config_log.txt 
+
+    FLAG_FILE="/tmp/first_boot_complete"
+
+    if [ ! -f "$FLAG_FILE" ]; then
+      # Redirect both stdout and stderr to the log file
+      exec > /tmp/install_log.txt 2>&1 
+
+      #sudo yum update -y #save time for first run
+      sudo yum install wget -y
+      sudo mkdir sw
+      sudo mkdir log
+      chmod 777 sw
+      chmod 777 log
+      cd sw
+      wget https://yum.oracle.com/repo/OracleLinux/OL9/appstream/x86_64/getPackage/oracle-database-preinstall-23ai-1.0-2.el9.x86_64.rpm
+      wget https://download.oracle.com/otn-pub/otn_software/db-free/oracle-database-free-23ai-1.0-1.el9.x86_64.rpm
+      sudo dnf -y install oracle-database-preinstall-23ai-1.0-2.el9.x86_64.rpm > /home/peter/log/install_log.txt
+      sudo dnf install -y oracle-database-free* >> /home/peter/log/install_log.txt
+      #export DB_PASSWORD=sys
+      (echo sys; echo sys;) | sudo /etc/init.d/oracle-free-23ai configure >> /home/peter/log/config_log.txt
+      # Create the flag file
+      touch "$FLAG_FILE"
+    fi 
   EOF
 }
